Implementation of search load service.

+++ TruckDepotDao/src/main/java/com/truckdepot/common/dao/repository/AddressRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.repository.CrudRepository;
 import com.truckdepot.common.dao.entities.Address;
 
 
-public interface AddressRepository extends CrudRepository<Address, Integer> {
+public interface AddressRepository extends CrudRepository<Address, Long> {
 	
-	Address findByAddressId(int addressId);
+	Address findByAddressId(long addressId);
 	
 
 }
