+++ TruckDepotDao/src/main/java/com/truckdepot/common/dao/repository/CarrierShipmentRepository.java
@@ -6,9 +6,9 @@ import com.truckdepot.common.dao.entities.CarrierShipment;
 import com.truckdepot.common.dao.entities.CarrierVehicle;
 
 
-public interface CarrierShipmentRepository extends CrudRepository<CarrierShipment, Integer> {
+public interface CarrierShipmentRepository extends CrudRepository<CarrierShipment, Long> {
 	
-	CarrierVehicle findById(int id);
+	CarrierVehicle findById(long id);
 	
 
 }
