+++ TruckDepotDao/src/main/java/com/truckdepot/common/dao/repository/QuotesPriceRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.repository.CrudRepository;
 import com.truckdepot.common.dao.entities.QuotesPrice;
 
 
-public interface QuotesPriceRepository extends CrudRepository<QuotesPrice, Integer> {
+public interface QuotesPriceRepository extends CrudRepository<QuotesPrice, Long> {
 	
-	QuotesPrice findById(int id);
+	QuotesPrice findById(long id);
 	
 
 }
