+++ TruckDepotDao/src/main/java/com/truckdepot/common/dao/repository/ShipperQuotesRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.repository.CrudRepository;
 import com.truckdepot.common.dao.entities.ShipperQuote;
 
 
-public interface ShipperQuotesRepository extends CrudRepository<ShipperQuote, Integer> {
+public interface ShipperQuotesRepository extends CrudRepository<ShipperQuote, Long> {
 	
-	ShipperQuote findByQuoteId(int quoteId);
+	ShipperQuote findByQuoteId(long quoteId);
 	
 
 }
