+++ TruckDepotDao/src/main/java/com/truckdepot/common/dao/repository/VehicleDriverRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.repository.CrudRepository;
 import com.truckdepot.common.dao.entities.CarrierVehicleDriver;
 
 
-public interface VehicleDriverRepository extends CrudRepository<CarrierVehicleDriver, Integer> {
+public interface VehicleDriverRepository extends CrudRepository<CarrierVehicleDriver, Long> {
 	
-	CarrierVehicleDriver findById(int id);
+	CarrierVehicleDriver findById(long id);
 	
 
 }
